requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-dotenv==1.0.1
Pillow==10.3.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
